automation-slots/requirements.txt
autopep8==1.6.0
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.4
easyocr==1.7.1
filelock==3.7.1
fsspec==2023.12.2
idna==3.6
imageio==2.33.1
Jinja2==3.1.2
keyboard==0.13.5
lazy_loader==0.3
Levenshtein==0.23.0
MarkupSafe==2.1.3
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
networkx==3.2.1
ninja==1.11.1.1
nodeenv==1.6.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.3.0
platformdirs==2.5.2
pyclipper==1.3.0.post5
pycodestyle==2.8.0
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-bidi==0.4.2
python-Levenshtein==0.23.0
pytweening==1.0.7
pywin32==306
PyYAML==6.0.1
rapidfuzz==3.6.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
sympy==1.12
tifffile==2023.12.9
toml==0.10.2
torch==2.1.2
torchvision==0.16.2
typing_extensions==4.9.0
unrpa==2.3.0
urllib3==2.1.0
virtualenv==20.14.1
